r1587: * Backportet find_next_bit for the x86-64 architecture. For   use of drbd-0.7.x with Linux-2.4.x on AMD's processors. * Made the warning in ppsize() on 64 architectures go away.

+++ drbd/drbd_fs.c
@@ -97,7 +97,7 @@ char* ppsize(char* buf, size_t size)
 		size = size >> 10;
 		base++;
 	}
-	sprintf(buf,"%d %cB",size,units[base]);
+	sprintf(buf,"%ld %cB",(long)size,units[base]);
 
 	return buf;
 }
